requirements.txt
flask
pytest
pytest-cov
pillow